Refactor WeakrefCallable to be a core handler
