+++ winforms/src/toga_winforms/app.py
@@ -10,10 +10,10 @@ from System.Media import SystemSounds
 from System.Net import SecurityProtocolType, ServicePointManager
 from System.Windows.Threading import Dispatcher
 
+from toga.handlers import WeakrefCallable
 from toga.dialogs import InfoDialog
 
 from .libs.proactor import WinformsProactorEventLoop
-from .libs.wrapper import WeakrefCallable
 from .screens import Screen as ScreenImpl
