+++ winforms/src/toga_winforms/dialogs.py
@@ -9,8 +9,9 @@ from System.Drawing import (
 )
 from System.Windows.Forms import DialogResult, MessageBoxButtons, MessageBoxIcon
 
+from toga.handlers import WeakrefCallable
+
 from .libs.user32 import DPI_AWARENESS_CONTEXT_UNAWARE, SetThreadDpiAwarenessContext
-from .libs.wrapper import WeakrefCallable
 
 
 class BaseDialog:
