+++ winforms/src/toga_winforms/widgets/button.py
@@ -4,8 +4,8 @@ import System.Windows.Forms as WinForms
 from travertino.size import at_least
 
 from toga.colors import TRANSPARENT
+from toga.handlers import WeakrefCallable
 
-from ..libs.wrapper import WeakrefCallable
 from .base import Widget
