+++ winforms/src/toga_winforms/widgets/canvas.py
@@ -24,10 +24,10 @@ from System.IO import MemoryStream
 
 from toga.colors import TRANSPARENT
 from toga.constants import Baseline, FillRule
+from toga.handlers import WeakrefCallable
 from toga.widgets.canvas import arc_to_bezier, sweepangle
 from toga_winforms.colors import native_color
 
-from ..libs.wrapper import WeakrefCallable
 from .box import Box
