+++ winforms/src/toga_winforms/widgets/dateinput.py
@@ -4,7 +4,8 @@ from decimal import ROUND_UP
 import System.Windows.Forms as WinForms
 from System import DateTime as WinDateTime
 
-from ..libs.wrapper import WeakrefCallable
+from toga.handlers import WeakrefCallable
+
 from .base import Widget
