+++ winforms/src/toga_winforms/widgets/multilinetextinput.py
@@ -2,10 +2,10 @@ import System.Windows.Forms as WinForms
 from System.Drawing import SystemColors
 from travertino.size import at_least
 
+from toga.handlers import WeakrefCallable
 from toga_winforms.colors import native_color
 from toga_winforms.libs.fonts import HorizontalTextAlignment
 
-from ..libs.wrapper import WeakrefCallable
 from .textinput import TextInput
