+++ winforms/src/toga_winforms/widgets/scrollcontainer.py
@@ -4,9 +4,9 @@ from System.Drawing import Point
 from System.Windows.Forms import Panel, SystemInformation
 from travertino.node import Node
 
+from toga.handlers import WeakrefCallable
 from toga_winforms.container import Container
 
-from ..libs.wrapper import WeakrefCallable
 from .base import Widget
 
 # On Windows, scroll bars usually appear only when the content is larger than the
