+++ winforms/src/toga_winforms/widgets/webview.py
@@ -13,6 +13,7 @@ from System.Drawing import Color
 from System.Threading.Tasks import Task, TaskScheduler
 
 import toga
+from toga.handlers import WeakrefCallable
 from toga.widgets.webview import CookiesResult, JavaScriptResult
 from toga_winforms.libs.extensions import (
     CoreWebView2Cookie,
@@ -21,7 +22,6 @@ from toga_winforms.libs.extensions import (
     WebView2RuntimeNotFoundException,
 )
 
-from ..libs.wrapper import WeakrefCallable
 from .base import Widget
